+++ src/main/java/com/funtravelapp/main/packageservice/service/impl/PackageServiceImpl.java
@@ -1,13 +1,11 @@
 package com.funtravelapp.main.packageservice.service.impl;
 
+import com.funtravelapp.main.authenticationservice.entity.User;
 import com.funtravelapp.main.packageservice.dto.PackageInputDTO;
 import com.funtravelapp.main.packageservice.entity.Package;
 import com.funtravelapp.main.packageservice.repository.PackageRepository;
 import com.funtravelapp.main.packageservice.service.PackageService;
-import com.funtravelapp.main.tokenAuth.dto.GetUserDTO;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -30,14 +28,18 @@ public class PackageServiceImpl implements PackageService {
 
 
     @Override
-    public Package insertNewPackage(String authorizationHeader, PackageInputDTO packageInputDTO) throws Exception {
-        if(packageInputDTO.getName().isBlank()){
-            throw new Exception("Please input package name");
+    public Package insertNewPackage(String authorizationHeader, Map<String, Boolean> role, User user, PackageInputDTO packageInputDTO) throws Exception {
+        if(packageInputDTO.getName().isBlank() || packageInputDTO.getDescription().isBlank() || packageInputDTO.getPrice().equals(0)){
+            throw new Exception("Please input blank fields!");
+        }
+
+        if (user.getRole().equalsIgnoreCase("customer")){
+            throw new Exception("Access denied!");
         }
 
         Package aPackage = new Package();
         aPackage.setId(0);
-        aPackage.setUserId(packageInputDTO.getUserId());
+        aPackage.setUserId(user.getId());
         aPackage.setName(packageInputDTO.getName());
         aPackage.setDescription(packageInputDTO.getDescription());
         aPackage.setPrice(packageInputDTO.getPrice());
@@ -46,31 +48,39 @@ public class PackageServiceImpl implements PackageService {
     }
 
     @Override
-    public List<Package> allPackages() {
+    public List<Package> allPackages(String authorizationHeader, Map<String, Boolean> role, User user) {
         List<Package> packages;
-        packages = repository.findAll();
-//        if (userDTO.getRole().equalsIgnoreCase("customer")){
-//            packages = repository.findAll();
-//        }else {
-//            packages = repository.findByUserId(userDTO.getId());
-//        }
+        System.out.println(user);
+        if (user.getRole().equalsIgnoreCase("customer")){
+            packages = repository.findAll();
+        }else {
+            packages = repository.findByUserId(user.getId());
+        }
         return packages;
     }
 
 
-    @Override
-    public List<Package> allPackageBySellerId(int sellerId) {
-        return repository.findByUserId(sellerId);
-    }
+//    @Override
+//    public List<Package> allPackageBySellerId(int sellerId) {
+//        return repository.findByUserId(sellerId);
+//    }
 
     @Override
-    public String uploadImage(MultipartFile file, Integer packageId) throws IOException {
+    public String uploadImage(String authorizationHeader, Map<String, Boolean> role, User user, Integer packageId, MultipartFile file) throws Exception {
+        if (user.getRole().equalsIgnoreCase("customer")){
+            throw new Exception("Access denied!");
+        }
+
         Optional<Package> getPackage = repository.findById(packageId);
 
         if(getPackage.isEmpty()){
             throw new IOException("400");
         }
         Package aPackage = getPackage.get();
+
+        if(!aPackage.getUserId().equals(user.getId())){
+            throw new Exception("Not allowed to access!");
+        }
 
         if(aPackage.getImage() != null){
             File existingFile = new File(folder + aPackage.getImage());
@@ -95,39 +105,59 @@ public class PackageServiceImpl implements PackageService {
     }
 
     @Override
-    public ResponseEntity<?> getPackageById(Integer packageId) {
-        return new ResponseEntity<>(repository.findPackageById(packageId), HttpStatus.OK );
+    public Package getPackageById(String authorizationHeader, Map<String, Boolean> role, User user, Integer packageId) {
+
+        return repository.findPackageById(packageId);
     }
 
     @Override
-    public ResponseEntity<String> delete(Integer id) {
+    public void delete(String authorizationHeader, Map<String, Boolean> role, User user, Integer id) throws Exception {
+
+        if (user.getRole().equalsIgnoreCase("customer")){
+            throw new Exception("Access denied!");
+        }
+
         Optional<Package> getPackage = repository.findById(id);
 
         if(getPackage.isEmpty()){
-            return new ResponseEntity<>("400", null);
+            throw new Exception("No package!");
         }
         Package aPackage = getPackage.get();
+
+        if(!aPackage.getUserId().equals(user.getId())){
+            throw new Exception("Not allowed to delete!");
+        }
 
         if(aPackage.getImage() != null){
             File existingFile = new File(folder + aPackage.getImage());
             existingFile.delete();
         }
         repository.delete(aPackage);
-
-        return new ResponseEntity<>("Deleted!", HttpStatus.OK);
     }
 
     @Override
-    public Package updatePackage(PackageInputDTO dto, Integer id) throws Exception {
+    public Package updatePackage(String authorizationHeader, Map<String, Boolean> role, User user, Integer id, PackageInputDTO dto) throws Exception {
         if(dto.getName().isBlank()){
             throw new Exception("Please input package name");
         }
 
-        Package aPackage = new Package();
-        aPackage.setId(id);
-        aPackage.setUserId(dto.getUserId());
+        if (user.getRole().equalsIgnoreCase("customer")){
+            throw new Exception("Access denied!");
+        }
+
+        Optional<Package> findPackage = repository.findById(id);
+
+        if(findPackage.isEmpty()){
+            throw new Exception("Not found!");
+        }
+
+        Package aPackage = findPackage.get();
+
+        if(!aPackage.getUserId().equals(user.getId())){
+            throw new Exception("Not allowed to edit!");
+        }
+
         aPackage.setName(dto.getName());
-        aPackage.setImage(dto.getImage());
         aPackage.setDescription(dto.getDescription());
         aPackage.setPrice(dto.getPrice());
